+++ src/components/Input/InputParts.module.css
@@ -58,7 +58,6 @@
 }
 
 .error-text {
-  margin-top: 5px;
   margin-left: 10px;
 
   font-size: var(--xm-size);
